backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "openai-agents[litellm]>=0.1.0",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.27.1",
    "sqlalchemy>=2.0.27",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.2.1",
    "email-validator>=1.3.0",
    "requests>=2.31.0",
    "pandas>=2.1.4",
    "numpy>=1.25.2",
    "scikit-learn>=1.3.2",
    "jose>=1.0.0",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",

]
